requirements.txt
streamlit
google-genai
groq
pydantic
langsmith
python-dotenv